gateway/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "archon-mcp-gateway"
version = "0.1.0"
description = "WS-6 MCP Host Gateway — authenticated proxy and plugin host for MCP servers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "PIE", "RET", "SIM"]
ignore = ["B008"]  # FastAPI Depends() in function signature default

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
